requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4
lxml
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Authlib==1.2.1
gunicorn==21.2.0
eventlet==0.33.0
python-socketio==5.10.0
Flask-SocketIO==5.3.0
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.26.0
tensorflow==2.14.0
keras==2.14.0
nltk==3.8.1
geopy==2.4.0
scipy==1.11.3
statsmodels==0.14.0
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-bootstrap-components==1.5.0
redis==5.0.1
celery==5.3.4
pymongo==4.6.0
dnspython==2.4.2
Werkzeug==2.3.7
itsdangerous==2.1.2
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
colorama==0.4.6
sendgrid==6.10.0